Refactor VirtualFileManager into ScopedFileManager, seperate files
