+++ apps/backend/src/models/file_manager/types.rs
@@ -1,0 +1,21 @@
+pub struct FSDirectoryEntry {
+	pub path: String,
+	pub entries: Vec<FSFileEntry>,
+}
+
+pub struct FSFileEntry {
+	pub path: String,
+	pub content: Vec<u8>,
+}
+
+pub enum FSEntry {
+	File(FSFileEntry),
+	Dir(FSDirectoryEntry),
+}
+
+pub enum FileManagerError {
+	NotFound,
+	PermissionDenied,
+	InvalidPath,
+	IoError(std::io::Error),
+}
