+++ apps/backend/src/models/mod.rs
@@ -1,6 +1,6 @@
+pub mod file_manager;
 pub mod file_schemas;
 pub mod game;
 pub mod hash;
 pub mod secrets;
 pub mod server;
-pub mod vfs;
